Move sendJsonResp function inside the renamed (tools -> helpers) helpers folder
