--- helper.js
@@ -1,2 +1,0 @@
-exports.sendJsonResp = (req, res) =>
-  res.json({ error: "", data: res.locals.response })
