+++ helpers/sendJsonResp.js
@@ -1,0 +1,4 @@
+const sendJsonResp = (req, res) =>
+  res.json({ error: "", data: res.locals.response })
+
+module.exports = sendJsonResp
